improved component styles
