+++ src/AwesomeApp.UI/package.json
@@ -22,6 +22,7 @@
     "eslint-config-next": "13.4.19",
     "iron-session": "^6.3.1",
     "next": "13.4.19",
+    "notistack": "^3.0.1",
     "postcss": "8.4.29",
     "react": "18.2.0",
     "react-dom": "18.2.0",
