+++ src/AwesomeApp.UI/src/pages/_components/spinner/styles.module.css
@@ -4,5 +4,6 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(240, 240, 240, 0.9);
+  background-color: rgba(255, 255, 255, 0.5);
+  z-index: 10;
 }
